+++ src/main/java/com/barco/admin/controller/AppUserRestApi.java
@@ -121,7 +121,7 @@ public class AppUserRestApi {
      * @param payload
      * @return ResponseEntity<?>
      * */
-    @PreAuthorize("hasRole('DEV')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(path="/deleteAllAppUserAccount", method=RequestMethod.POST)
     public ResponseEntity<?> deleteAllAppUserAccount(@RequestBody AppUserRequest payload) {
         try {
@@ -137,7 +137,7 @@ public class AppUserRestApi {
      * @apiNote :- Api use to download app user account template file
      * @return ResponseEntity<?> downloadAppUserAccountTemplateFile
      * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(value = "/downloadAppUserAccountTemplateFile", method = RequestMethod.GET)
     public ResponseEntity<?> downloadAppUserAccountTemplateFile() {
         try {
@@ -157,7 +157,7 @@ public class AppUserRestApi {
      * @apiNote :- Api use to download app user account
      * @return ResponseEntity<?> AppUserRequest
      * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(value = "/downloadAppUserAccount", method = RequestMethod.POST)
     public ResponseEntity<?> downloadAppUserAccount(@RequestBody AppUserRequest payload) {
         try {
@@ -178,7 +178,7 @@ public class AppUserRestApi {
      * @param payload
      * @return ResponseEntity<?>
      * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(value = "/fetchAllAppUserAccount", method = RequestMethod.POST)
     public ResponseEntity<?> fetchAllAppUserAccount(@RequestBody AppUserRequest payload) {
         try {
@@ -195,7 +195,7 @@ public class AppUserRestApi {
      * @param payload
      * @return ResponseEntity<?>
      * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(value = "/addAppUserAccount", method = RequestMethod.POST)
     public ResponseEntity<?> addAppUserAccount(@RequestBody AppUserRequest payload) {
         try {
@@ -215,7 +215,7 @@ public class AppUserRestApi {
      * @param payload
      * @return ResponseEntity<?>
      * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(value = "/editAppUserAccount", method = RequestMethod.POST)
     public ResponseEntity<?> editAppUserAccount(@RequestBody AppUserRequest payload) {
         try {
@@ -235,7 +235,7 @@ public class AppUserRestApi {
      * @param payload
      * @return ResponseEntity<?>
      * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('DEV')")
     @RequestMapping(value = "/enabledDisabledAppUserAccount", method = RequestMethod.POST)
     public ResponseEntity<?> enabledDisabledAppUserAccount(@RequestBody AppUserRequest payload) {
         try {
