+++ src/main/java/com/barco/admin/controller/FormSettingRestApi.java
@@ -37,142 +37,6 @@ public class FormSettingRestApi {
     private FormSettingService formSettingService;
 
     /**
-     * @apiName :- addSTT
-     * @apiNote :- Api use to create stt (source task type)
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/addSTT", method = RequestMethod.POST)
-    public ResponseEntity<?> addSTT(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.addSTT(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while addSTT ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- editSTT
-     * @apiNote :- Api use to update stt (source task type)
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/editSTT", method = RequestMethod.POST)
-    public ResponseEntity<?> editSTT(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.editSTT(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while editSTT ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- deleteSTT
-     * @apiNote :- Api use to delete stt (source task type)
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/deleteSTT", method = RequestMethod.POST)
-    public ResponseEntity<?> deleteSTT(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.deleteSTT(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while deleteSTT ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- fetchSTTBySttId
-     * @apiNote :- Api use to fetch stt by stt id(source task type)
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/fetchSTTBySttId", method = RequestMethod.POST)
-    public ResponseEntity<?> fetchSTTBySttId(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.fetchSTTBySttId(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while fetchSTTBySttId ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- fetchAllSTT
-     * @apiNote :- Api use to fetch stt(source task type)
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/fetchAllSTT", method = RequestMethod.POST)
-    public ResponseEntity<?> fetchAllSTT(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.fetchAllSTT(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while fetchAllSTT ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- deleteAllSTT
-     * @apiNote :- Api use to delete all stt
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/deleteAllSTT", method = RequestMethod.POST)
-    public ResponseEntity<?> deleteAllSTT(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.deleteAllSTT(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while deleteAllSTT ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- fetchAllSTTLinkForm
-     * @apiNote :- Api use to fetch link stt with form
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/fetchAllSTTLinkForm", method = RequestMethod.POST)
-    public ResponseEntity<?> fetchAllSTTLinkForm(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.fetchAllSTTLinkForm(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while fetchAllSTTLinkForm ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * @apiName :- linkSTTForm
-     * @apiNote :- Api use to link stt with form
-     * @param payload
-     * @return ResponseEntity<?>
-     * */
-    @PreAuthorize("hasRole('MASTER_ADMIN') or hasRole('ADMIN') or hasRole('USER')")
-    @RequestMapping(value = "/linkSTTForm", method = RequestMethod.POST)
-    public ResponseEntity<?> linkSTTForm(@RequestBody STTRequest payload) {
-        try {
-            return new ResponseEntity<>(this.formSettingService.linkSTTForm(payload), HttpStatus.OK);
-        } catch (Exception ex) {
-            logger.error("An error occurred while linkSTTForm ", ExceptionUtil.getRootCause(ex));
-            return new ResponseEntity<>(new AppResponse(BarcoUtil.ERROR, ex.getMessage()), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
      * @apiName :- addForm
      * @apiNote :- Api use to add form(source form)
      * @param payload
@@ -241,7 +105,7 @@ public class FormSettingRestApi {
     }
 
     /**
-     * @apiName :- fetchSTTF
+     * @apiName :- fetchForms
      * @apiNote :- Api use to fetch form
      * @param payload
      * @return ResponseEntity<?>
