+++ src/main/java/com/barco/admin/service/impl/QueryService.java
@@ -74,7 +74,7 @@ public class QueryService {
     public static String DELETE_APP_USER_EVENT_BRIDGE_BY_EVENT_BRIDGE_ID_AND_APP_USER_ID = "DELETE FROM APP_USER_EVENT_BRIDGE AUEB WHERE EVENT_BRIDGE_ID = %d AND APP_USER_ID = %d  ";
     public static String DELETE_APP_USER_PROFILE_ACCESS_BY_ROLE_ID_AND_APP_USER_ID = "DELETE FROM APP_USER_PROFILE_ACCESS aupa WHERE aupa.PROFILE_ID = %d AND aupa.APP_USER_ID = %d ";
     public static String FETCH_LINK_ROLE_WITH_ROOT_USER = "SELECT DISTINCT AU.ID, AU.EMAIL, AU.USERNAME, AU.FIRST_NAME || ' ' || AU.LAST_NAME AS FULL_NAME, " +
-        "AU.IMG AS PROFIlE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, AURA.DATE_CREATED AS LINK_DATA, " +
+        "AU.IMG AS PROFIlE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, PRO.DESCRIPTION, AURA.DATE_CREATED AS LINK_DATA, " +
         "CASE WHEN AURA.DATE_CREATED IS NULL THEN FALSE ELSE TRUE END LINKED, " +
         "CASE WHEN AURA.STATUS IS NOT NULL THEN AURA.STATUS WHEN AU.STATUS = 0 OR RL.STATUS = 0 THEN 0 ELSE 1 END AS LINK_STATUS " +
         "FROM APP_USER AU " +
@@ -84,7 +84,7 @@ public class QueryService {
         "WHERE AU.STATUS != 2 AND (AU.ID = %d OR AU.CREATED_BY_ID = %d) " +
         "ORDER BY AU.ID DESC";
     public static String FETCH_LINK_PROFILE_WITH_ROOT_USER = "SELECT DISTINCT AU.ID, AU.EMAIL, AU.USERNAME, AU.FIRST_NAME || ' ' || AU.LAST_NAME AS FULL_NAME, " +
-        "AU.IMG AS PROFIlE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, AUPA.DATE_CREATED AS LINK_DATA, " +
+        "AU.IMG AS PROFIlE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, PRO.DESCRIPTION, AUPA.DATE_CREATED AS LINK_DATA, " +
         "CASE WHEN AUPA.DATE_CREATED IS NULL THEN FALSE ELSE TRUE END LINKED, " +
         "CASE WHEN AUPA.STATUS IS NOT NULL THEN AUPA.STATUS WHEN AU.STATUS = 0 OR PRO.STATUS = 0 THEN 0 ELSE 1 END AS LINK_STATUS " +
         "FROM APP_USER AU " +
@@ -93,7 +93,7 @@ public class QueryService {
         "WHERE AU.STATUS != 2 AND (AU.ID = %d OR AU.CREATED_BY_ID = %d) " +
         "ORDER BY AU.ID DESC";
     public static String FETCH_LINK_ENVIRONMENT_VARIABLE_WITH_USER = "SELECT DISTINCT AU.ID, AU.EMAIL, AU.USERNAME, AU.FIRST_NAME || ' ' || AU.LAST_NAME AS FULL_NAME, " +
-        "AU.IMG AS PROFILE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, AUE.DATE_CREATED AS LINK_DATA, " +
+        "AU.IMG AS PROFILE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, PRO.DESCRIPTION, AUE.DATE_CREATED AS LINK_DATA, " +
         "CASE WHEN AUE.DATE_CREATED IS NULL THEN FALSE ELSE TRUE END LINKED, AUE.ENV_VALUE, " +
         "CASE WHEN AUE.STATUS IS NOT NULL THEN AUE.STATUS WHEN AU.STATUS = 0 OR AUE.STATUS = 0 THEN 0 ELSE 1 END AS LINK_STATUS " +
         "FROM APP_USER AU " +
@@ -102,7 +102,7 @@ public class QueryService {
         "WHERE AU.STATUS != %d " +
         "ORDER by AU.ID DESC";
     public static String FETCH_LINK_EVENT_BRIDGE_WITH_USER = "SELECT DISTINCT AU.ID, AU.EMAIL, AU.USERNAME, AU.FIRST_NAME || ' ' || AU.LAST_NAME AS FULL_NAME, " +
-        "AU.IMG AS PROFILE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, AUEB.DATE_CREATED AS LINK_DATA, " +
+        "AU.IMG AS PROFILE_IMG, PRO.ID AS PROFILE_ID, PRO.PROFILE_NAME, PRO.DESCRIPTION, AUEB.DATE_CREATED AS LINK_DATA, " +
         "CASE WHEN AUEB.DATE_CREATED IS NULL THEN FALSE ELSE TRUE END LINKED, AUEB.ACCESS_TOKEN,  AUEB.EXPIRE_TIME,  AUEB.TOKEN_ID, " +
         "CASE WHEN AUEB.STATUS IS NOT NULL THEN AUEB.STATUS WHEN AU.STATUS = 0 OR AUEB.STATUS = 0 THEN 0 ELSE 1 END AS LINK_STATUS " +
         "FROM APP_USER AU " +
